LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/pseudo-elements-007-expected.txt


PASS Initial font-size for ::before
PASS Initial font-size for ::after
PASS Initial font-size for ::marker
PASS Initial font-size for ::first-line
PASS Initial font-size for ::first-letter
PASS Initial font-size for ::backdrop
FAIL font-size for ::before depending on container assert_equals: expected "30px" but got "80px"
FAIL font-size for ::after depending on container assert_equals: expected "30px" but got "80px"
FAIL font-size for ::marker depending on container assert_equals: expected "30px" but got "80px"
FAIL font-size for ::first-line depending on container assert_equals: expected "30px" but got "80px"
FAIL font-size for ::first-letter depending on container assert_equals: expected "30px" but got "80px"
FAIL font-size for ::backdrop depending on container assert_equals: expected "30px" but got "80px"
